backend/requirements.txt
fastapi
uvicorn[standard]
pymysql
mysql-connector-python
pyspark
